refactor: 테스트 도메인명 Ai로 변경
#36
